% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcreportRTMB-package.R
\docType{package}
\name{mcreportRTMB-package}
\alias{mcreportRTMB}
\alias{mcreportRTMB-package}
\title{mcreportRTMB: Interface with RTMB to take posterior samples of derived quantities.}
\description{
Interface with RTMB to take posterior samples of derived quantities.
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
