% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcreport.R
\name{MCREPORT}
\alias{MCREPORT}
\title{Make a quantity in an RTMB objective function available to mcreport}
\usage{
MCREPORT(x)
}
\arguments{
\item{x}{The quantity to take samples of}
}
\value{
Used for its side effects when running mcreport
}
\description{
Make a quantity in an RTMB objective function available to mcreport
}
